refactor: Section 관리 기능을 LineService로 이동
